heatmap

+++ eacirc/circuit/gate2/CpuGate.cpp
@@ -5,7 +5,7 @@
 void CpuGate::run( const Node* circ, const TestVectors& ins, TestVectors& outs )
 {
     Interpreter interpreter( &spec_, layers_.get() );
-    for ( int i = 0; i < ins.num(); ++i ) {
+    for ( size_t i = 0; i < ins.num(); ++i ) {
         interpreter.execute( circ, ins[i], outs[i] );
     }
 }
